My first attempt at a v3 ballbot using stepping motors.
Unfortunately, I discovered that the stepping motors I am using cannot drive
the omniwheels with enough torque when the omniwheel are being forced to
slip sideways. I attempted this with two different omniwheels and both
failed. I believe my next step is to use geared DC motors, but that will
require redesign.

Even thought this code was not able to be made to work I am checking it in
as a starting place for some future ballbot. Be warned that it is in some
intermediate debugging state.